test: create blackbox test for HTTP Handlers
Move tests that evaluate public functions into a separate file
repackaged with the new name http_test. This offers blackbox testing.
